printGraph method added + miscellaneous bug fixes
I came up with a way to print the representation that we currently have.
If we change the representation, we'll probably have to change this.  I
still am not sure that we want to use three array lists, but it was the
only solution that came to mind.  It's probably efficient in terms of
time?  Maybe we should brainstorm more about this.

renamed vertexes to vertices because i nerd

Added reciprocity to the addEdge method.

Added a check for a null pointer in the addEdge method, when a pointer
goes through a for loop to find the end of a particular vertex's linked
list.

Oh, and I realized that using "|" as a delimiter doesn't work unless you
type the pattern as "\\|" , because | means "or" when you're writing out
a pattern.  So using "|" as your delimiter pattern translates to
"anything OR anything" (I think).