Vertex lists are now sorted
Adding an edge to a linked list at a vertex is now done so that the
values are stored from greatest to least.

(It's easy to reverse this if we want to.)

Added some comments for clarification.

Started shortestChain.  I'm currently thinking it should be a boolean,
so this method can double as a way to find whether somebody is
connected.  We could alternatively have it return some other data type
that might be more useful, or "null" when it doesn't find any
connection.  Perhaps an array of the names traversed in the shortest
path?