(Students in a School) method done
I'm still sure we need to improve efficiency, but I wrote a method for
finding the students in a school and generalized printGraph so that it
would work for the subgraph generated by school().  I also fixed the
addVertex method, because it wasn't adding the working before.  The
problem was that I was using == instead of .equals().

Also made "print" an option.

Started to draft the shortestChain method, barely.  It doesn't do
anything whatsoever yet.  I think this was covered in recitation, but I
wasn't paying attention...